rag/requirements.txt
# RAG Server Dependencies
# Install in a venv: python -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt
# Note: Uses LanceDB instead of ChromaDB for Python 3.14 compatibility

lancedb>=0.4.0
pandas>=2.0.0
sentence-transformers>=2.2.0
einops>=0.6.0
langchain-text-splitters>=0.0.1
httpx>=0.25.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pyyaml>=6.0
